Tentative login and register screen is made
