+++ reader-app/src/screens/LoginScreen.tsx
@@ -1,0 +1,81 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, TouchableOpacity, StyleSheet } from 'react-native';
+
+export default function LoginScreen({ navigation }: any) {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleLogin = () => {
+    // İleride buraya veritabanı (backend) kontrolü gelecek
+    console.log('Sgining:', email);
+    // Şimdilik direkt hikayeye yönlendiriyoruz
+    navigation.navigate('Story');
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Tekrar Hoş Geldiniz</Text>
+
+      <TextInput
+        style={styles.input}
+        placeholder="E-mail"
+        placeholderTextColor="#888"
+        value={email}
+        onChangeText={setEmail}
+        keyboardType="email-address"
+        autoCapitalize="none"
+      />
+
+      <TextInput
+        style={styles.input}
+        placeholder="Şifreniz"
+        placeholderTextColor="#888"
+        value={password}
+        onChangeText={setPassword}
+        secureTextEntry // Şifreyi yıldızlı gösterir
+      />
+
+      <TouchableOpacity style={styles.button} onPress={handleLogin}>
+        <Text style={styles.buttonText}>Giriş Yap</Text>
+      </TouchableOpacity>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    padding: 20,
+    backgroundColor: '#121212',
+  },
+  title: {
+    color: '#FFFFFF',
+    fontSize: 28,
+    fontWeight: 'bold',
+    marginBottom: 40,
+    textAlign: 'center',
+  },
+  input: {
+    backgroundColor: '#1E1E1E',
+    color: '#FFFFFF',
+    padding: 15,
+    borderRadius: 8,
+    marginBottom: 15,
+    borderWidth: 1,
+    borderColor: '#333',
+    fontSize: 16,
+  },
+  button: {
+    backgroundColor: '#BB86FC',
+    padding: 15,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginTop: 10,
+  },
+  buttonText: {
+    color: '#121212',
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+});
